Print integer arrays accomplished.
A function to print an integer array in format […, …]
I used to finish it in Java. So I think I should do such a thing in C as well.
I also did a test in the main.c. And print the elements in arrays. Moreover, the function will return a string, or namely a char array.