A complex set operations has been finished.
I implement two operations of set via tree in array, including FindR, which is used to find the root node index in the array, and Union, which is used to make union of two set.
Set is an important data structure that can be implemented by tree. So I do these.
Warning! This is not the best way to store the data. There is still a better way.