The best set operations has been finished.
I implement two operations of set via tree in array, including FindR, which is used to find the root node index in the array, and Union, which is used to make union of two set.
Set is an important data structure that can be implemented by tree. So I do these.
This edition uses half the memory of the complex edition, which is 4 bytes now and 8 bytes before.